Improve backtrace
Unfortunately, this is still not a good solution. It's hard to combine the
mechanisms of `exec_fun_indirect` and `next`. They change the flow of Rust
code, so Rust's call stack becomes inadequate. Ideally, everything should be
done through `next` and Forth's return stack, so that we can simply traverse
the return stack.